+++ pkg/network/server_config.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/nspcc-dev/neo-go/pkg/config"
 	"github.com/nspcc-dev/neo-go/pkg/config/netmode"
-	"github.com/nspcc-dev/neo-go/pkg/wallet"
 	"go.uber.org/zap/zapcore"
 )
 
@@ -63,7 +62,7 @@ type (
 		LogLevel zapcore.Level
 
 		// Wallet is a wallet configuration.
-		Wallet *wallet.Config
+		Wallet *config.Wallet
 
 		// TimePerBlock is an interval which should pass between two successive blocks.
 		TimePerBlock time.Duration
@@ -76,7 +75,7 @@ func NewServerConfig(cfg config.Config) ServerConfig {
 	appConfig := cfg.ApplicationConfiguration
 	protoConfig := cfg.ProtocolConfiguration
 
-	var wc *wallet.Config
+	var wc *config.Wallet
 	if appConfig.UnlockWallet.Path != "" {
 		wc = &appConfig.UnlockWallet
 	}
